AG-8006 - Rename file
